requirements.txt
torch
torchvision
transformers
pillow
kornia
setproctitle
pytest
pytest-timeout
ruff
einops
timm
dazpycheck
coverage
